+++ include/engine/vector.h
@@ -1,18 +1,18 @@
 #ifndef VECTOR_H
 #define VECTOR_H
 
-#include <engine/types.h>
+#include <stddef.h>
 
 typedef struct {
   void *items;
-  usize item_size;
-  usize len;
-  usize capacity;
+  size_t item_size;
+  size_t len;
+  size_t capacity;
 } vector_t;
 
-vector_t *vector_create(usize item_size, usize capacity);
-u8 vector_append(vector_t *vector, void *item, usize *out_index);
-void *vector_get(vector_t *vector, usize index);
-u8 vector_remove(vector_t *vector, usize index);
+vector_t *vector_create(size_t item_size, size_t capacity);
+int vector_append(vector_t *vector, void *item, size_t *out_index);
+void *vector_get(vector_t *vector, size_t index);
+int vector_remove(vector_t *vector, size_t index);
 
 #endif
